[fix] fixed NewSchema ignores opts arg
